+++ src/Publish/Console/PublishCommand.php
@@ -9,15 +9,12 @@ declare(strict_types=1);
 
 namespace Ixocreate\Application\Publish\Console;
 
+use FilesystemIterator;
 use Ixocreate\Application\Console\CommandInterface;
 use Ixocreate\Application\Publish\PublishConfig;
-use Ixocreate\Application\Publish\PublishDefinitionConfig;
 use Ixocreate\Filesystem\Storage\StorageSubManager;
-use League\Flysystem\Adapter\Local;
-use League\Flysystem\FileNotFoundException;
-use League\Flysystem\Filesystem;
-use League\Flysystem\FilesystemInterface;
-use League\Flysystem\MountManager;
+use RecursiveDirectoryIterator;
+use RecursiveIteratorIterator;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
@@ -26,11 +23,6 @@ use Symfony\Component\Console\Output\OutputInterface;
 
 final class PublishCommand extends Command implements CommandInterface
 {
-    /**
-     * @var PublishDefinitionConfig
-     */
-    private $publishDefinitionConfig;
-
     /**
      * @var PublishConfig
      */
@@ -41,11 +33,10 @@ final class PublishCommand extends Command implements CommandInterface
      */
     private $storageSubManager;
 
-    public function __construct(PublishDefinitionConfig $publishDefinitionConfig, PublishConfig $publishConfig, StorageSubManager $storageSubManager)
+    public function __construct(PublishConfig $publishConfig, StorageSubManager $storageSubManager)
     {
         parent::__construct(self::getCommandName());
 
-        $this->publishDefinitionConfig = $publishDefinitionConfig;
         $this->publishConfig = $publishConfig;
         $this->storageSubManager = $storageSubManager;
     }
@@ -63,55 +54,80 @@ final class PublishCommand extends Command implements CommandInterface
             ->setAliases(['publish']);
     }
 
+    /**
+     * @param InputInterface $input
+     * @param OutputInterface $output
+     * @return int|void|null
+     * @throws \Exception
+     */
     public function execute(InputInterface $input, OutputInterface $output)
     {
-        $definitions = $this->publishDefinitionConfig->getDefinitions($this->publishConfig);
+        $definitions = $this->publishConfig->getDefinitions();
 
-        if (!empty($input->getArgument("type"))) {
+        if (!empty($input->getArgument('type'))) {
             foreach ($definitions as $name => $definition) {
-                if ($name === $input->getArgument("type")) {
+                if ($name === $input->getArgument('type')) {
                     $definitions[$name] = $definition;
                     break;
                 }
             }
         }
 
-        $rootStorage = new Filesystem(new Local(\getcwd(), LOCK_EX, Local::SKIP_LINKS));
-
         foreach ($definitions as $name => $definition) {
-            /** @var FilesystemInterface $storage */
-            $storage = $this->storageSubManager->get($definition['storage']);
 
-            $mountManager = new MountManager([
-                'root' => $rootStorage,
-                'storage' => $storage,
-            ]);
+            if (!\is_dir($definition['targetDirectory'])) {
+                throw new \Exception('target is not a directory: ' . $definition['targetDirectory']);
+            }
+            if (!\is_writable($definition['targetDirectory'])) {
+                throw new \Exception('target is not writable: ' . $definition['targetDirectory']);
+            }
+
+            $targetPermissions = \fileperms($definition['targetDirectory']);
 
             foreach ($definition['sources'] as $directory) {
-                foreach ($mountManager->listContents('root://' . $directory, true) as $content) {
-                    if ($content['type'] !== "file") {
+
+                $iterator = new RecursiveIteratorIterator(
+                    new RecursiveDirectoryIterator($directory, FilesystemIterator::SKIP_DOTS),
+                    RecursiveIteratorIterator::SELF_FIRST
+                );
+
+                foreach ($iterator as $file) {
+                    /** @var \SplFileInfo $file */
+                    $path = $file->getPathname();
+
+                    if (\preg_match('#(^|/|\\\\)\.{1,2}$#', $path)) {
                         continue;
                     }
 
-                    $sourceFile = \str_replace($directory . "/", "", $content['path']);
-
-                    if ((bool) $input->getOption("force") === false && $mountManager->has('storage://' . $sourceFile)) {
+                    if (!$file->isFile()) {
                         continue;
                     }
 
-                    try {
-                        $deleted = $mountManager->delete('storage://' . $sourceFile);
-                    } catch (FileNotFoundException $e) {
-                        $deleted = true;
+                    $sourceFile = \str_replace($directory . '/', '', $file->getPathname());
+                    $targetFile = $definition['targetDirectory'] . $sourceFile;
+
+                    if (\file_exists($targetFile)) {
+                        if ((bool) $input->getOption('force') === false) {
+                            continue;
+                        }
+                        \unlink($targetFile);
                     }
 
-                    if ($deleted) {
-                        $mountManager->copy('root://' . $content['path'], 'storage://' . $sourceFile);
+                    $targetDirectory = \dirname($targetFile);
+
+                    if (!\is_dir($targetDirectory)) {
+                        if (!@\mkdir($targetDirectory, $targetPermissions, true)) {
+                            throw new \Exception(error_get_last());
+                        }
                     }
+
+                    \copy($path, $targetFile);
+
+                    continue;
                 }
             }
 
-            $output->writeln(\sprintf("%s published", $name));
+            $output->writeln(\sprintf('%s published', $name));
         }
     }
 }