+++ CHANGELOG.md
@@ -2,8 +2,12 @@
 
 ## [Unreleased](https://github.com/ixocreate/application/compare/0.1.0...develop)
 
+### Changed
+- merge PublishDefinitionConfig with PublishConfig
+- remove Filesystem dependency
+
 ### Fixed
-- Changelog Compare Link
+- Changelog Compare Lin
 
 ## [v0.1.0 (2019-04-19)](https://github.com/ixocreate/application/compare/master...0.1.0)